adding package.json and package-lock.json moving reused props to propDefinition
