+++ components/kanbanize/actions/create-new-task/create-new-task.mjs
@@ -1,5 +1,4 @@
 import kanbanizeApp from "../../kanbanize.app.mjs";
-import constants from "../../common/constants.mjs";
 
 export default {
   key: "kanbanize-create-new-task",
@@ -10,53 +9,44 @@ export default {
   props: {
     kanbanizeApp,
     boardId: {
-      label: "Board Id",
-      description: "The ID of the board to move the task into.\n\nYou can see the board ID on the dashboard screen, in the upper right corner of each board.",
-      type: "string",
-      reloadProps: true,
-      async options() {
-        return this.kanbanizeApp.getBoardsOptions();
-      },
+      propDefinition: [
+        kanbanizeApp,
+        "boardId",
+      ],
     },
     title: {
-      label: "Title",
-      description: "Title of the task.",
-      type: "string",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "title",
+      ],
     },
     description: {
-      label: "Description",
-      description: "Description of the task.",
-      type: "string",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "description",
+      ],
     },
     type: {
-      label: "Type",
-      description: "The type of the task.",
-      type: "string",
-      optional: true,
-      async options() {
-        if (!this.boardId) {
-          return [];
-        }
-        return this.kanbanizeApp.getTypes(this.boardId);
-      },
+      propDefinition: [
+        kanbanizeApp,
+        "type",
+        (c) => ({
+          boardId: c.boardId,
+        }),
+      ],
     },
     assignee: {
-      label: "Assignee",
-      description: "Username of the assignee.",
-      type: "string",
-      optional: true,
-      async options() {
-        if (!this.boardId) {
-          return [];
-        }
-        return this.kanbanizeApp.getUsernames(this.boardId);
-      },
+      propDefinition: [
+        kanbanizeApp,
+        "assignee",
+        (c) => ({
+          boardId: c.boardId,
+        }),
+      ],
     },
     column: {
       label: "Column",
-      description: "Get only cards from a specific column.\n\nThis field has higher priority than the section option.",
+      description: "The name of the column to create the task into (default is Backlog).",
       type: "string",
       optional: true,
       async options() {
@@ -68,7 +58,7 @@ export default {
     },
     lane: {
       label: "Lane",
-      description: "Only get cards from that specific lane.",
+      description: "The name of the swimlane to put the newly created task into.\n\nIf omitted, the task will be placed in the first swimlane.",
       type: "string",
       optional: true,
       async options() {
@@ -103,41 +93,40 @@ export default {
       optional: true,
     },
     priority: {
-      label: "Priority",
-      description: "Priority of the task.",
-      type: "string",
-      optional: true,
-      options: constants.priorityOpts,
+      propDefinition: [
+        kanbanizeApp,
+        "priority",
+      ],
     },
     size: {
-      label: "Size",
-      description: "Size of the task.",
-      type: "integer",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "size",
+      ],
     },
     extLink: {
-      label: "External Link",
-      description: "A external link in the following format: `https://domain.com/resource`",
-      type: "string",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "extLink",
+      ],
     },
     deadline: {
-      label: "Deadline",
-      description: "Deadline in the format: yyyy-mm-dd (e.g. 2011-12-13).",
-      type: "string",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "deadline",
+      ],
     },
     color: {
-      label: "Color",
-      description: "Any color code (e.g. #34A97B).",
-      type: "string",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "color",
+      ],
     },
     tags: {
-      label: "Tags",
-      description: "List of tags.",
-      type: "string[]",
-      optional: true,
+      propDefinition: [
+        kanbanizeApp,
+        "tags",
+      ],
     },
   },
   async additionalProps() {
