+++ components/kanbanize/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@pipedream/kanbanize",
+    "version": "0.0.1",
+    "description": "Pipedream Kanbanize Components",
+    "main": "kanbanize.app.mjs",
+    "keywords": [
+        "pipedream",
+        "kanbanize"
+    ],
+    "homepage": "https://pipedream.com/apps/kanbanize",
+    "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
+    "license": "MIT",
+    "gitHead": "e12480b94cc03bed4808ebc6b13e7fdb3a1ba535",
+    "publishConfig": {
+        "access": "public"
+    },
+    "dependencies": {
+        "@pipedream/platform": "^0.10.0"
+    }
+}
